+++ tst/OpcUaStackClient/ValueBasedInterface/VBIAsyncReal_Attribute_t.cpp
@@ -4,8 +4,6 @@
 
 using namespace OpcUaStackClient;
 
-#ifdef REAL_SERVER
-
 BOOST_AUTO_TEST_SUITE(VBIAsyncReal_Attribute_)
 
 BOOST_AUTO_TEST_CASE(VBIAsyncReal_Attribute_)
@@ -13,10 +11,47 @@ BOOST_AUTO_TEST_CASE(VBIAsyncReal_Attribute_)
 	std::cout << "VBIAsyncReal_Attribute_t" << std::endl;
 }
 
+#ifdef REAL_SERVER
 BOOST_AUTO_TEST_CASE(VBIAsyncReal_Attribute_read)
 {
 	VBIClientHandlerTest vbiClientHandlerTest;
 	VBIClient client;
+
+
+	//
+	// init certificate and crypto manager
+	//
+	ApplicationCertificate::SPtr applicationCertificate = constructSPtr<ApplicationCertificate>();
+	applicationCertificate->enable(true);
+
+	applicationCertificate->certificateTrustListLocation("./pki/trusted/certs/");
+	applicationCertificate->certificateRejectListLocation("./pki/reject/certs/.");
+	applicationCertificate->certificateRevocationListLocation("./pki/trusted/crl/");
+	applicationCertificate->issuersCertificatesLocation("./pki/issuers/certs/");
+	applicationCertificate->issuersRevocationListLocation("./pki/issuers/crl/");
+
+	applicationCertificate->serverCertificateFile("./pki/own/certs/ASNeG-Demo.der");
+	applicationCertificate->privateKeyFile("./pki/own/private/ASNeG-Demo.pem");
+
+	applicationCertificate->generateCertificate(true);
+	applicationCertificate->uri("urn:asneg.de:ASNeG:ASNeG-Demo");
+	applicationCertificate->commonName("ASNeG-Demo");
+	applicationCertificate->domainComponent("127.0.0.1");
+	applicationCertificate->organization("ASNeG");
+	applicationCertificate->organizationUnit("OPC UA Service Department");
+	applicationCertificate->locality("Neukirchen");
+	applicationCertificate->state("Hessen");
+	applicationCertificate->country("DE");
+	applicationCertificate->yearsValidFor(5);
+	applicationCertificate->keyLength(2048);
+	applicationCertificate->certificateType("RsaSha256");
+	applicationCertificate->ipAddress().push_back("127.0.0.1");
+	applicationCertificate->dnsName().push_back("ASNeG.de");
+	applicationCertificate->email("info@ASNeG.de");
+
+	BOOST_REQUIRE(applicationCertificate->init() == true);
+	CryptoManager::SPtr cryptoManager = constructSPtr<CryptoManager>();
+
 
 	// set session change callback
 	client.setSessionChangeCallback(
@@ -27,6 +62,9 @@ BOOST_AUTO_TEST_CASE(VBIAsyncReal_Attribute_read)
 	ConnectContext connectContext;
 	connectContext.endpointUrl_ = REAL_SERVER_URI;
 	connectContext.sessionName_ = REAL_SESSION_NAME;
+	connectContext.applicationCertificate_ = applicationCertificate;
+	connectContext.cryptoManager_ = cryptoManager;
+	connectContext.secureChannelLog_ = true;
 	vbiClientHandlerTest.sessionStateUpdate_.initEvent();
 	client.asyncConnect(connectContext);
 	BOOST_REQUIRE(vbiClientHandlerTest.sessionStateUpdate_.waitForEvent(1000) == true);
@@ -58,6 +96,42 @@ BOOST_AUTO_TEST_CASE(VBIAsyncReal_Attribute_write)
 	VBIClientHandlerTest vbiClientHandlerTest;
 	VBIClient client;
 
+
+	//
+	// init certificate and crypto manager
+	//
+	ApplicationCertificate::SPtr applicationCertificate = constructSPtr<ApplicationCertificate>();
+	applicationCertificate->enable(true);
+
+	applicationCertificate->certificateTrustListLocation("./pki/trusted/certs/");
+	applicationCertificate->certificateRejectListLocation("./pki/reject/certs/.");
+	applicationCertificate->certificateRevocationListLocation("./pki/trusted/crl/");
+	applicationCertificate->issuersCertificatesLocation("./pki/issuers/certs/");
+	applicationCertificate->issuersRevocationListLocation("./pki/issuers/crl/");
+
+	applicationCertificate->serverCertificateFile("./pki/own/certs/ASNeG-Demo.der");
+	applicationCertificate->privateKeyFile("./pki/own/private/ASNeG-Demo.pem");
+
+	applicationCertificate->generateCertificate(true);
+	applicationCertificate->uri("urn:asneg.de:ASNeG:ASNeG-Demo");
+	applicationCertificate->commonName("ASNeG-Demo");
+	applicationCertificate->domainComponent("127.0.0.1");
+	applicationCertificate->organization("ASNeG");
+	applicationCertificate->organizationUnit("OPC UA Service Department");
+	applicationCertificate->locality("Neukirchen");
+	applicationCertificate->state("Hessen");
+	applicationCertificate->country("DE");
+	applicationCertificate->yearsValidFor(5);
+	applicationCertificate->keyLength(2048);
+	applicationCertificate->certificateType("RsaSha256");
+	applicationCertificate->ipAddress().push_back("127.0.0.1");
+	applicationCertificate->dnsName().push_back("ASNeG.de");
+	applicationCertificate->email("info@ASNeG.de");
+
+	BOOST_REQUIRE(applicationCertificate->init() == true);
+	CryptoManager::SPtr cryptoManager = constructSPtr<CryptoManager>();
+
+
 	// set session change callback
 	client.setSessionChangeCallback(
 		boost::bind(&VBIClientHandlerTest::sessionStateUpdate, &vbiClientHandlerTest, (uint32_t)1234, _1, _2)
@@ -67,6 +141,9 @@ BOOST_AUTO_TEST_CASE(VBIAsyncReal_Attribute_write)
 	ConnectContext connectContext;
 	connectContext.endpointUrl_ = REAL_SERVER_URI;
 	connectContext.sessionName_ = REAL_SESSION_NAME;
+	connectContext.applicationCertificate_ = applicationCertificate;
+	connectContext.cryptoManager_ = cryptoManager;
+	connectContext.secureChannelLog_ = true;
 	vbiClientHandlerTest.sessionStateUpdate_.initEvent();
 	client.asyncConnect(connectContext);
 	BOOST_REQUIRE(vbiClientHandlerTest.sessionStateUpdate_.waitForEvent(1000) == true);
@@ -77,7 +154,7 @@ BOOST_AUTO_TEST_CASE(VBIAsyncReal_Attribute_write)
 	OpcUaNodeId nodeId;
 	OpcUaDataValue dataValue;
 	dataValue.variant()->set((OpcUaBoolean)1);
-	nodeId.set("Demo.Static.Scalar.Boolean", 2);
+	nodeId.set(220, 1);
 	vbiClientHandlerTest.writeComplete_.initEvent();
 	client.asyncWrite(
 		nodeId,
@@ -94,7 +171,6 @@ BOOST_AUTO_TEST_CASE(VBIAsyncReal_Attribute_write)
 	BOOST_REQUIRE(vbiClientHandlerTest.sessionState_ == SS_Disconnect);
 	BOOST_REQUIRE(vbiClientHandlerTest.clientHandle_ == 1234);
 }
+#endif
 
 BOOST_AUTO_TEST_SUITE_END()
-
-#endif